renamed client to consumer
