--- samples/commandline/client/src/main/java/org/eclipse/dataspaceconnector/client/command/azure/vault/AzureVaultGetSecretRequestExecutor.java
@@ -1,34 +1,0 @@
-/*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.client.command.azure.vault;
-
-import org.eclipse.dataspaceconnector.client.command.CommandExecutor;
-import org.eclipse.dataspaceconnector.client.command.CommandResult;
-import org.eclipse.dataspaceconnector.client.command.ExecutionContext;
-
-import static org.eclipse.dataspaceconnector.client.command.http.HttpOperations.executeGet;
-
-public class AzureVaultGetSecretRequestExecutor implements CommandExecutor {
-    @Override
-    public CommandResult execute(ExecutionContext context) {
-        var key = context.getParams().stream().findFirst();
-
-        if (key.isPresent()) {
-            return executeGet("/api/vault?key="+key.get(), context);
-        }
-
-        throw new IllegalArgumentException("needs at least one parameter!");
-    }
-}
