--- samples/commandline/client/src/main/java/org/eclipse/dataspaceconnector/client/command/ids/IdsConstants.java
@@ -1,26 +1,0 @@
-/*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.client.command.ids;
-
-import java.net.URI;
-
-public interface IdsConstants {
-
-    String ID = "DAG-X_CLIENT";
-
-    URI ID_URI = URI.create("DAG-X_CLIENT");
-
-    String VERSION = "1.0";
-}
