+++ samples/commandline/consumer/src/main/java/org/eclipse/dataspaceconnector/consumer/command/CommandExecutor.java
@@ -1,0 +1,30 @@
+/*
+ *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.consumer.command;
+
+/**
+ * Executes a consumer command.
+ */
+@FunctionalInterface
+public interface CommandExecutor {
+
+    /**
+     * Executes the command.
+     *
+     * @param context the execution context
+     */
+    CommandResult execute(ExecutionContext context);
+
+}
