+++ samples/commandline/consumer/src/main/java/org/eclipse/dataspaceconnector/consumer/common/Dates.java
@@ -1,0 +1,38 @@
+/*
+ *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.consumer.common;
+
+import javax.xml.datatype.DatatypeConfigurationException;
+import javax.xml.datatype.DatatypeFactory;
+import javax.xml.datatype.XMLGregorianCalendar;
+import java.util.Date;
+import java.util.GregorianCalendar;
+
+public class Dates {
+
+    /**
+     * Returns a calendar representing the current instant.
+     */
+    public static XMLGregorianCalendar gregorianNow() {
+        try {
+            GregorianCalendar calendar = new GregorianCalendar();
+            calendar.setTime(new Date());
+            return DatatypeFactory.newInstance().newXMLGregorianCalendar(calendar);
+        } catch (DatatypeConfigurationException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+}
